backend/requirements.txt
Django
djangorestframework
torch
psycopg2-binary
django-cors-headers
openai-whisper